WaE vs2015: pointer to int casting
Reviewed-on: https://gerrit.libreoffice.org/22592
Tested-by: Jenkins <ci@libreoffice.org>
(cherry picked from commit 3496535a3754438f314b7b143e7223b498ca3ec9)

+++ sal/osl/w32/random.c
@@ -16,7 +16,7 @@ int osl_get_system_random_data(char* buffer, size_t desired_len)
     unsigned int val;
 
     /* if unaligned fill to alignment */
-    if((int)buffer & 3)
+    if((uintptr_t)buffer & 3)
     {
         size_t len = 4 - ((size_t)(buffer) & 3);
